pyproject.toml
[project]
name = "test-auth"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
]
